fix: MV3 utils
