+++ app/src/features/rules/modals/ImportRulesModal/actions/index.js
@@ -7,7 +7,7 @@ import { runRuleMigrations } from "utils/rules/ruleMigrations";
 import APP_CONSTANTS from "config/constants";
 import { RecordStatus } from "features/rules";
 import { isExtensionManifestVersion3 } from "actions/ExtensionActions";
-import { migrateRuleToMV3 } from "modules/extension/ruleParser";
+import { migrateRuleToMV3 } from "modules/extension/utils";
 //CONSTANTS
 const { RULES_LIST_TABLE_CONSTANTS } = APP_CONSTANTS;
