+++ app/src/modules/extension/mv3RuleParser/parseHeadersRule.ts
