+++ app/src/modules/extension/mv3RuleParser/parseRedirectRule.ts
