+++ app/src/modules/extension/mv3RuleParser/parseReplaceRule.ts
