+++ app/src/modules/extension/mv3RuleParser/parseScriptRule.ts
