+++ app/src/modules/extension/mv3RuleParser/parseUserAgentRule.ts
