+++ app/src/modules/extension/mv3RuleParser/utils.ts
